// player/types.d.ts
interface CSSTransformSpecialParam<T> {
	x: number;
	y: T;
	dx: T;
	dy: T;
	movex: T;
	movey: T;
}

export interface Style
	extends CSS.Properties<string | number>,
		CSS.PropertiesHyphen<string | number>,
		Partial<CSSTransformSpecialParam<number>> {}

export interface StyleEntry {
	[x: string]:
		| string
		| number
		| { value: number; unit?: string; ease?: string; original: string }
		| LerpStringStyle;
}

export interface LerpStringStyle {
	value: Array<number>;
	pattern: Array<string>;
	ease?: string;
	original: string;
}
// ...
export type Broadcast =
	| BroadcastAction
	| {
			type?: BroadcastAction;
			transition?: Transition;
			volume?: number;
			playbackRate?: number;
	  };
// ...
export interface Transition {
	from?: Record<string, any>;
	to: Record<string, any>;
	duration?: number;
	repeat?: number;
	yoyo?: boolean;
	onComplete?: () => void;
	ease?: string | Array<string | Record<string, string>>;
}
// ...
interface PersoBase extends Perso {
	parent: string;
	node: HTMLElement | SVGElement;
	style: Style;
	transform: TransformProperty;
	matrix: Matrix2D | [];
	content: Set<PersoId>;
}

export interface PersoText extends PersoBase {
	type: PersoType.TEXT;
	child: Txt;
}

export interface PersoLayer extends PersoBase {
	type: PersoType.LAYER;
	child: Layer;
}

export interface PersoImg extends PersoBase {
	type: PersoType.IMG;
	initial: Partial<Initial> & Img;
	update: (update: Partial<Initial> & Img) => void;
}

export interface PersoSprite extends PersoBase {
	type: PersoType.SPRITE;
	initial: Partial<Initial> & Img;
	update: (update: Partial<Initial> & Img) => void;
}

export interface PersoVideo extends PersoBase {
	type: PersoType.VIDEO;
	initial: Partial<Initial> & Vid;
	media: HTMLVideoElement | HTMLAudioElement;
	update: (update: Partial<Broadcast>) => void;
}

export interface PersoSound extends PersoBase {
	type: PersoType.SOUND;
	initial: Partial<Initial> & Vid;
	media: My;
	update: (update: Partial<Broadcast>) => void;
}

export interface SoundNode extends PersoSoundDef {
	media?: My;
}

export interface My extends MediaElementAudioSourceNode {
	my?: {
		connect: () => void;
		disconnect: () => void;
	};
}

export type PersoNode = PersoText | PersoLayer | PersoSprite | PersoImg | PersoSound | PersoVideo;

export interface Initial {
	tag: string;
	id: string;
	attr: any;
	style: Style;
	classStyle: Style;
	className?: string | ActionClassList;
	move: string;
}

export interface Vid {
	src: string;
}

// A revoir passer les props en style sauf src et img
export interface Img {
	img?: HTMLImageElement;
	src: string;
	fit?: string;
	ratio?: number;
	width?: number | string;
	height?: number | string;
	px?: string;
	py?: string;
}
export type ImagesCollection = Map<string, Img>;

export enum PersoType {
	TEXT = 'text',
	IMG = 'img',
	LIST = 'list',
	BLOC = 'bloc',
	ROOT = 'root',
	VIDEO = 'video',
	PROTO = 'proto',
	LAYER = 'layer',
	SPRITE = 'sprite',
	BUTTON = 'button',
	POLYGON = 'polygon',
	SOUND = 'sound',
	AUDIO = 'audio',
}